Wrap define_controller initializer in to_prepare
This was causing problems when I attempted to upgrade to Rails 7: my
`ApplicationController` references some code that doesn't get loaded
until Rails has been initialized, and this engine was attempting to load
`ApplicationController` before Rails is initialized. This resulted in
the code not being found. Wrapping it in a `to_prepare` block fixed the
problem.

This also fixes a Rubocop being disabled and not re-enabled afterward.